+++ common/src/main/java/net/SpectrumFATM/black_archive/blockentity/door/PillarDoorModel.java
@@ -19,7 +19,10 @@ public class PillarDoorModel extends ShellDoorModel {
 	public static TexturedModelData getTexturedModelData() {
 		ModelData modelData = new ModelData();
 		ModelPartData modelPartData = modelData.getRoot();
-		ModelPartData main = modelPartData.addChild("main", ModelPartBuilder.create().uv(0, 3).cuboid(0.0F, -15.0F, 13.0F, 14.0F, 32.0F, 0.0F, new Dilation(0.0F)), ModelTransform.pivot(-7.0F, 7.0F, -5.0F));
+		ModelPartData main = modelPartData.addChild("main", ModelPartBuilder.create().uv(0, 3).cuboid(0.0F, -15.0F, 13.0F, 14.0F, 32.0F, 0.0F, new Dilation(0.0F))
+				.uv(0, 35).cuboid(-3.0F, -18.0F, 12.0F, 3.0F, 35.0F, 3.0F, new Dilation(0.0F))
+				.uv(1, 35).cuboid(14.0F, -18.0F, 12.0F, 3.0F, 35.0F, 3.0F, new Dilation(0.0F))
+				.uv(60, 0).cuboid(0.0F, -17.0F, 12.0F, 14.0F, 2.0F, 3.0F, new Dilation(0.0F)), ModelTransform.pivot(-7.0F, 7.0F, -5.0F));
 
 		ModelPartData door = modelPartData.addChild("door", ModelPartBuilder.create().uv(28, 1).cuboid(0.0F, -15.0F, -2.1F, 14.0F, 32.0F, 2.0F, new Dilation(0.0F)), ModelTransform.pivot(-7.0F, 7.0F, 10.0F));
 		ShellModel.addMaterializationPart(modelPartData);