+++ common/src/main/java/net/SpectrumFATM/black_archive/entity/client/ModModelLayers.java
@@ -1,20 +1,15 @@
 package net.SpectrumFATM.black_archive.entity.client;
 
-import net.SpectrumFATM.black_archive.blockentity.ModShells;
-import net.SpectrumFATM.black_archive.blockentity.door.PillarDoorBlock;
+import net.SpectrumFATM.black_archive.blockentity.door.PillarDoorModel;
 import net.SpectrumFATM.black_archive.blockentity.shell.PillarShellModel;
-import net.fabricmc.fabric.api.client.rendering.v1.EntityModelLayerRegistry;
-import net.minecraft.client.model.ModelPartBuilder;
-import net.minecraft.client.render.block.entity.BlockEntityRenderer;
 import net.minecraft.client.render.entity.model.EntityModelLayer;
 import net.minecraft.client.render.entity.model.EntityModelLoader;
-import net.minecraft.client.render.entity.model.EntityModels;
 import net.minecraft.util.Identifier;
 
 public class ModModelLayers {
 
     public static PillarShellModel PILLAR_EXT;
-    public static PillarDoorBlock PILLAR_DOOR;
+    public static PillarDoorModel PILLAR_DOOR;
 
     public static final EntityModelLayer DALEK = new EntityModelLayer(new Identifier("black_archive", "dalek"), "main");
     public static final EntityModelLayer DALEK_SLAVE = new EntityModelLayer(new Identifier("black_archive", "dalek_slave"), "main");
@@ -27,6 +22,6 @@ public class ModModelLayers {
 
     public static void setupModels(EntityModelLoader entityModels) {
         PILLAR_EXT = new PillarShellModel(entityModels.getModelPart(PILLAR_SHELL_LAYER));
-        PILLAR_DOOR = new PillarDoorBlock(entityModels.getModelPart(PILLAR_DOOR_LAYER));
+        PILLAR_DOOR = new PillarDoorModel(entityModels.getModelPart(PILLAR_DOOR_LAYER));
     }
 }