+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Forge Support!!!**
 - Tissue Compression Eliminator
+- Pillar Exterior
 
 ### Improved
 - Time rotor now continues to movie when in temporal orbit.